prevent div by zero in bitrate calc in case of stupid truncation of small duration allow optional output of all qualities in one fragmented MP4
git-svn-id: https://mp4parser.googlecode.com/svn/trunk/isoparser@461 7decde4b-c250-0410-a0da-51896bc88be6

+++ src/main/java/com/googlecode/mp4parser/authoring/builder/smoothstreaming/FlatManifestWriterImpl.java
@@ -177,7 +177,7 @@ public class FlatManifestWriterImpl implements ManifestWriter {
             bitrate += sample.limit();
         }
         bitrate *= 8; // from bytes to bits
-        bitrate /= getDuration(track) / track.getTrackMetaData().getTimescale(); // per second
+        bitrate /= ((double)getDuration(track)) / track.getTrackMetaData().getTimescale(); // per second
         return bitrate;
     }
